requirements.txt
airbyte-cdk==1.5.1
airbyte_protocol_models==0.14.1
airbyte_protocol_models_dataclasses==0.14.1
annotated-types==0.7.0
anyio==4.8.0
attributes-doc==0.4.0
attrs==24.3.0
backoff==2.2.1
bracex==2.5.post1
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==42.0.8
Deprecated==1.2.15
dpath==2.2.0
dunamai==1.23.0
genson==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.6.1
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==0.2
jsonschema==3.2.0
langchain-core==0.1.42
langsmith==0.1.147
MarkupSafe==3.0.2
nltk==3.9.1
numpy==1.26.4
orjson==3.10.14
packaging==23.2
pandas==2.2.2
pendulum==2.1.2
platformdirs==4.3.6
psutil==6.1.0
pycparser==2.22
pydantic==1.10.21
pydantic_core==2.27.2
PyJWT==2.10.1
pyrate-limiter==3.1.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==3.0.0
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-toolbelt==1.0.0
serpyco-rs==1.11.0
six==1.17.0
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
url-normalize==1.4.3
urllib3==2.3.0
wcmatch==8.4
wrapt==1.17.1
xmltodict==0.13.0
